OPDIS : Fixed type errors in test programs

+++ test/disasm_bfd.c
@@ -11,8 +11,8 @@
 
 static int print_insn( opdis_insn_t * insn, void * arg ) {
 	const char * filename = (const char *) arg;
-	printf( "%08X [%s:%X]\t%s\n", insn->vma, filename, insn->offset,
-		insn->ascii );
+	printf( "%08X [%s:%X]\t%s\n", (unsigned int) insn->vma, filename, 
+		(unsigned int) insn->offset, insn->ascii );
 	return 1;
 }
